Create README.md
Time Series Forecasting (Multivariate Regression) - This was an independent project idea that was carried out with another team member. We evaluated what economic and political factors can be good predictors of the Unemployment Rate. In addition to evaluating economic indicators such as Consumer Price Index, Consumer Sentiment, Consumer Confidence, Trade Balance, Strength of the Dollar and Oil Prices, we also wanted to analyze how much impact the current Presidential Party in office has on the level of Unemployment. Furthermore, we evaluated the impact of GDP on the Unemployment Rate using the economic theory of Okun's Law. Key tasks of the project included testing the model for assumptions of a strong linear representation and correcting the model for the presence of autocorrelation in a time-series analysis.